# Fix `alpha-value-notation` performance with improved benchmark script