--- a/.changeset/tricky-gorillas-drive.md
+++ b/.changeset/tricky-gorillas-drive.md
@@ -1,0 +1,5 @@
+---
+"stylelint": patch
+---
+
+Fixed: `alpha-value-notation` performance with improved benchmark script