--- a/lib/rules/alpha-value-notation/index.js
+++ b/lib/rules/alpha-value-notation/index.js
@@ -23,27 +23,11 @@ const meta = {
 	fixable: true,
 };
 
-const ALPHA_PROPS = new Set([
-	'opacity',
-	'shape-image-threshold',
-	// SVG properties
-	'fill-opacity',
-	'flood-opacity',
-	'stop-opacity',
-	'stroke-opacity',
-]);
-const ALPHA_FUNCS = new Set([
-	'color',
-	'hsl',
-	'hsla',
-	'hwb',
-	'lab',
-	'lch',
-	'oklab',
-	'oklch',
-	'rgb',
-	'rgba',
-]);
+const ALPHA_PROPS_REGEXP =
+	/^(?:opacity|shape-image-threshold|fill-opacity|flood-opacity|stop-opacity|stroke-opacity)$/i;
+const ALPHA_FUNCTION_REGEXP = /(?:color|hsla?|rgba?|hwb|lab|lch|oklab|oklch)\(/i;
+const ALPHA_FUNCTION_NAME_REGXP = /^(?:color|hsla?|rgba?|hwb|lab|lch|oklab|oklch)$/i;
+const DIGIT_REGEXP = /\d/;
 
 /** @type {import('stylelint').Rule} */
 const rule = (primary, secondaryOptions, context) => {
@@ -78,20 +62,31 @@ const rule = (primary, secondaryOptions, context) => {
 		});
 
 		root.walkDecls((decl) => {
+			const declarationValue = getDeclarationValue(decl);
+			const isAlphaProp = ALPHA_PROPS_REGEXP.test(decl.prop);
+
+			if (
+				// If the property is not an alpha property
+				(!isAlphaProp &&
+					// And the value is not an alpha function
+					!ALPHA_FUNCTION_REGEXP.test(declarationValue)) ||
+				// Or the value does not contain digits
+				!DIGIT_REGEXP.test(declarationValue)
+			) {
+				// Abort early
+				return;
+			}
+
 			let needsFix = false;
-			const parsedValue = valueParser(getDeclarationValue(decl));
+			const parsedValue = valueParser(declarationValue);
 
 			parsedValue.walk((node) => {
 				/** @type {import('postcss-value-parser').Node | undefined} */
 				let alpha;
 
-				if (ALPHA_PROPS.has(decl.prop.toLowerCase())) {
+				if (isAlphaProp && DIGIT_REGEXP.test(node.value)) {
 					alpha = findAlphaInValue(node);
-				} else {
-					if (node.type !== 'function') return;
-
-					if (!ALPHA_FUNCS.has(node.value.toLowerCase())) return;
-
+				} else if (node.type === 'function' && ALPHA_FUNCTION_NAME_REGXP.test(node.value)) {
 					alpha = findAlphaInFunction(node);
 				}