--- a/lib/rules/alpha-value-notation/index.js
+++ b/lib/rules/alpha-value-notation/index.js
@@ -26,7 +26,7 @@ const meta = {
 const ALPHA_PROPS_REGEXP =
 	/^(?:opacity|shape-image-threshold|fill-opacity|flood-opacity|stop-opacity|stroke-opacity)$/i;
 const ALPHA_FUNCTION_REGEXP = /(?:color|hsla?|rgba?|hwb|lab|lch|oklab|oklch)\(/i;
-const ALPHA_FUNCTION_NAME_REGXP = /^(?:color|hsla?|rgba?|hwb|lab|lch|oklab|oklch)$/i;
+const ALPHA_FUNCTION_NAME_REGEXP = /^(?:color|hsla?|rgba?|hwb|lab|lch|oklab|oklch)$/i;
 const DIGIT_REGEXP = /\d/;
 
 /** @type {import('stylelint').Rule} */