--- a/lib/rules/alpha-value-notation/index.js
+++ b/lib/rules/alpha-value-notation/index.js
@@ -86,7 +86,7 @@ const rule = (primary, secondaryOptions, context) => {
 
 				if (isAlphaProp && DIGIT_REGEXP.test(node.value)) {
 					alpha = findAlphaInValue(node);
-				} else if (node.type === 'function' && ALPHA_FUNCTION_NAME_REGXP.test(node.value)) {
+				} else if (node.type === 'function' && ALPHA_FUNCTION_NAME_REGEXP.test(node.value)) {
 					alpha = findAlphaInFunction(node);
 				}