--- a/lib/rules/alpha-value-notation/index.js
+++ b/lib/rules/alpha-value-notation/index.js
@@ -23,11 +23,11 @@ const meta = {
 	fixable: true,
 };
 
-const ALPHA_PROPS_REGEXP =
+const ALPHA_PROPS =
 	/^(?:opacity|shape-image-threshold|fill-opacity|flood-opacity|stop-opacity|stroke-opacity)$/i;
-const ALPHA_FUNCTION_REGEXP = /(?:color|hsla?|rgba?|hwb|lab|lch|oklab|oklch)\(/i;
-const ALPHA_FUNCTION_NAME_REGEXP = /^(?:color|hsla?|rgba?|hwb|lab|lch|oklab|oklch)$/i;
-const DIGIT_REGEXP = /\d/;
+const ALPHA_FUNCTION = /(?:color|hsla?|rgba?|hwb|lab|lch|oklab|oklch)\(/i;
+const ALPHA_FUNCTION_NAME = /^(?:color|hsla?|rgba?|hwb|lab|lch|oklab|oklch)$/i;
+const DIGIT = /\d/;
 
 /** @type {import('stylelint').Rule} */
 const rule = (primary, secondaryOptions, context) => {
@@ -63,15 +63,15 @@ const rule = (primary, secondaryOptions, context) => {
 
 		root.walkDecls((decl) => {
 			const declarationValue = getDeclarationValue(decl);
-			const isAlphaProp = ALPHA_PROPS_REGEXP.test(decl.prop);
+			const isAlphaProp = ALPHA_PROPS.test(decl.prop);
 
 			if (
 				// If the property is not an alpha property
 				(!isAlphaProp &&
 					// And the value is not an alpha function
-					!ALPHA_FUNCTION_REGEXP.test(declarationValue)) ||
+					!ALPHA_FUNCTION.test(declarationValue)) ||
 				// Or the value does not contain digits
-				!DIGIT_REGEXP.test(declarationValue)
+				!DIGIT.test(declarationValue)
 			) {
 				// Abort early
 				return;
@@ -84,9 +84,9 @@ const rule = (primary, secondaryOptions, context) => {
 				/** @type {import('postcss-value-parser').Node | undefined} */
 				let alpha;
 
-				if (isAlphaProp && DIGIT_REGEXP.test(node.value)) {
+				if (isAlphaProp && DIGIT.test(node.value)) {
 					alpha = findAlphaInValue(node);
-				} else if (node.type === 'function' && ALPHA_FUNCTION_NAME_REGEXP.test(node.value)) {
+				} else if (node.type === 'function' && ALPHA_FUNCTION_NAME.test(node.value)) {
 					alpha = findAlphaInFunction(node);
 				}