--- a/scripts/benchmark-rule.mjs
+++ b/scripts/benchmark-rule.mjs
@@ -53,7 +53,7 @@ fetch(CSS_URL)
 	.then((response) => {
 		let css = '';
 
-		for (let i = 0; i < 10; i++) {
+		for (let i = 0; i < 20; i++) {
 			css += `${response}\n\n`;
 		}