<!-- Each pull request must be associated with an open issue unless it's a documentation fix. If a corresponding issue does not exist, please create one so we can discuss the change first. -->

<!-- Please answer the following. We close pull requests that don't. -->

> Which issue, if any, is this issue related to?

See : https://github.com/stylelint/stylelint/issues/6853

> Is there anything in the PR that needs further explanation?

I was curious about performance and if there were any benchmark tools in place.

I found that the current benchmark script is just checking PostCSS performance, not rule performance in specific. I've refactored this script to separate the initial parsing phase and the plugin processing.

This gives a more relevant result to Stylelint plugins.

I also found that processing Bootstrap only once doesn't give meaningful results.
PostCSS is just too fast, the deviation too high and the actual duration per cycle with or without a code change isn't clear.

I found that duplicating the Bootstrap source 20 times gives a more useful output.


```
> node scripts/benchmark-rule.mjs alpha-value-notation ["number"]

Warnings: 0
Mean: 94.41161338888888 ms
Deviation: 12.21578916466738 ms
```

```
> node scripts/benchmark-rule.mjs alpha-value-notation ["number"]

Warnings: 0
Mean: 72.47793288888887 ms
Deviation: 6.865530836418062 ms
```

I have a M2 macbook air, which is possibly the worst machine to run benchmarks on.
So someone with a decent machine (no power saving cores, cooled cpu) should verify this work.

-------

This change also adds some improvements to the first rule.
I just picked the first rule at random, not because I noticed anything wrong in terms of performance.

- add a fast abort before any parsing
- some minor code improvements

--------

Changes like these do not take away the underlying issue as described in https://github.com/stylelint/stylelint/issues/6853

But I do think it is worthwhile to look into and apply trivial improvements like these.
Even with a different style parser it will be a good practice to add fast aborts before doing a deeper AST walk.